Make drbd do autodiscovery via exported resources.
This commit replaces the `host1`/`host2` and `ip1`/`ip2` parameters with
a `cluster` parameter. The configuration file is created via concat
fragments, and the drbd cluster is initialized when both the primary and
secondary resources have been realized.

The service resource has been make virtual and moved into its own class
so that dependencies may be placed on it, but it may not be realized in
the case that both primary & secondary resources are not present.

This commit also adds a `mountpoint` parameter such that an alternate
mountpoint may be specified for the drbd-resource.

This change is backwards-incompatible.