Create drbd::resource::up for a dependency target
When using the drbd module, the exported resources and `if defined(...)`
ordering is a bit complicated. This commit creates a
`Drbd::Resource::Up[$resource]` target for dependencies that will only
exist if the `defined()` function passes. This allows for services such
as mysql and corosync to delay their setup until DRBD is set up.