fix metadisk parameter
in commit aad63cc6cc995745f4ac3440a9654e256bd473be
the meta-disk switched to flexible-meta-disk.
For internal metadisk, the two options are equal, but for
external disk this can brak the configuration.

This commit adds the possibility to configure a flexible-metadisk
by not breaking existing configs.

This commit is backwards compatible with the last tagged version 0.5.2
but breaks compatibility with actual master.